refactor: extract chain type to common/chain subpackage
Move Chain type and constants to common/chain to avoid pulling in
heavy utils.go dependencies (echo, xz, protobuf) when only the chain
type is needed. common/chain.go re-exports via type aliases for
backwards compatibility. Saves ~387KB.
