+++ address/address.go
@@ -5,17 +5,17 @@ import (
 
 	"github.com/vultisig/mobile-tss-lib/tss"
 
-	"github.com/vultisig/vultisig-go/common"
+	"github.com/vultisig/vultisig-go/common/chain"
 )
 
 // GetAddress returns the address, public key, isEdDSA, and error for the given public key and chain.
-func GetAddress(rootHexPublicKey string, rootChainCode string, chain common.Chain) (address string, publicKey string, isEdDSA bool, err error) {
+func GetAddress(rootHexPublicKey string, rootChainCode string, c chain.Chain) (address string, publicKey string, isEdDSA bool, err error) {
 	if len(rootHexPublicKey) != 66 && len(rootHexPublicKey) != 64 {
 		return "", "", false, fmt.Errorf("invalid public key: %s", rootHexPublicKey)
 	}
 
-	if !chain.IsEdDSA() {
-		publicKey, err = tss.GetDerivedPubKey(rootHexPublicKey, rootChainCode, chain.GetDerivePath(), chain.IsEdDSA())
+	if !c.IsEdDSA() {
+		publicKey, err = tss.GetDerivedPubKey(rootHexPublicKey, rootChainCode, c.GetDerivePath(), c.IsEdDSA())
 		if err != nil {
 			return "", "", false, fmt.Errorf("failed to derive public key: %w", err)
 		}
@@ -23,80 +23,80 @@ func GetAddress(rootHexPublicKey string, rootChainCode string, chain common.Chai
 		publicKey = rootHexPublicKey
 	}
 
-	switch chain {
-	case common.Bitcoin:
+	switch c {
+	case chain.Bitcoin:
 		address, err = GetBitcoinAddress(publicKey)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.BitcoinCash:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.BitcoinCash:
 		address, err = GetBitcoinCashAddress(publicKey)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Litecoin:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Litecoin:
 		address, err = GetLitecoinAddress(publicKey)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Dogecoin:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Dogecoin:
 		address, err = GetDogecoinAddress(publicKey)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.GaiaChain:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.GaiaChain:
 		address, err = GetBech32Address(publicKey, `cosmos`)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.THORChain:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.THORChain:
 		address, err = GetBech32Address(publicKey, `thor`)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.MayaChain:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.MayaChain:
 		address, err = GetBech32Address(publicKey, `maya`)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Kujira:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Kujira:
 		address, err = GetBech32Address(publicKey, `kujira`)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Dydx:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Dydx:
 		address, err = GetBech32Address(publicKey, `dydx`)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.TerraClassic, common.Terra:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.TerraClassic, chain.Terra:
 		address, err = GetBech32Address(publicKey, `terra`)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Osmosis:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Osmosis:
 		address, err = GetBech32Address(publicKey, `osmo`)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Noble:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Noble:
 		address, err = GetBech32Address(publicKey, `noble`)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Akash:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Akash:
 		address, err = GetBech32Address(publicKey, `akash`)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Arbitrum, common.Base, common.BscChain, common.Ethereum, common.Polygon, common.Blast, common.Avalanche, common.Optimism, common.CronosChain, common.Zksync, common.Mantle, common.Hyperliquid, common.Sei:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Arbitrum, chain.Base, chain.BscChain, chain.Ethereum, chain.Polygon, chain.Blast, chain.Avalanche, chain.Optimism, chain.CronosChain, chain.Zksync, chain.Mantle, chain.Hyperliquid, chain.Sei:
 		address, err = GetEVMAddress(publicKey)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Sui:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Sui:
 		address, err = GetSuiAddress(publicKey)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Solana:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Solana:
 		address, err = GetSolAddress(publicKey)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Zcash:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Zcash:
 		address, err = GetZcashAddress(publicKey)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Dash:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Dash:
 		address, err = GetDashAddress(publicKey)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Tron:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Tron:
 		address, err = GetTronAddress(publicKey)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.XRP:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.XRP:
 		address, err = GetXRPAddress(publicKey)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Polkadot:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Polkadot:
 		address, err = GetDotAddress(publicKey)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Bittensor:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Bittensor:
 		address, err = GetBittensorAddress(publicKey)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Cardano:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Cardano:
 		address, err = GetCardanoAddress(publicKey)
-		return address, publicKey, chain.IsEdDSA(), err
-	case common.Ton:
+		return address, publicKey, c.IsEdDSA(), err
+	case chain.Ton:
 		address, err = GetTonAddress(publicKey)
-		return address, publicKey, chain.IsEdDSA(), err
+		return address, publicKey, c.IsEdDSA(), err
 	default:
-		return "", "", false, fmt.Errorf("unsupported chain: %s", chain)
+		return "", "", false, fmt.Errorf("unsupported chain: %s", c)
 	}
 }
